feat: adicionando nova arquitetura | corrigindo instabilidade | adicionando arquivo Makefile ao inves dos scripts individuais
