+++ main.c
@@ -6,23 +6,22 @@
 #include "./models/usuario.c"
 
 #if defined(__MINGW32__) || defined(_MSC_VER)
-#define limpar_input() fflush(stdin)
 #define limpar_tela() system("cls")
 #else
-#include "./include/stdio_ext.h"
-#define limpar_input() __fpurge(stdin)
 #define limpar_tela() system("clear")
 #endif
 
-void main(void) {
+void main(void){
   // criando diretorio de storage
   system("mkdir storage");
   limpar_tela();
   // checando se existe um usuário cadastrado.
   int result = user_validate();
 
-  if(result == 0)
-    return user_create();
-    
-  return menu();
+  if(result == 0){
+    user_create();
+  }
+  else{
+    menu();
+  }     
 }
